여정 여행 삽입 정상화

+++ member-service/src/main/java/com/traplaner/memberservice/member/service/MemberService.java
@@ -82,7 +82,7 @@ public class MemberService {
         return member;
     }
 
-    public Member findById(long id) {
+    public Member findById(int id) {
         return memberRepository.findById(id).orElseThrow(
                 ()->new EntityNotFoundException("Member not found")
         );
