+++ travelboard-service/src/main/java/com/traplaner/travelboardservice/travelBoard/repository/FavoriteRepository.java
@@ -1,8 +1,8 @@
 package com.traplaner.travelboardservice.travelBoard.repository;
 
-import com.traplaner.travelBoard.entity.Favorite;
+import com.traplaner.travelboardservice.travelBoard.entity.Favorite;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
+public interface FavoriteRepository extends JpaRepository<Favorite, Integer> {
 
 }
