+++ travelboard-service/src/main/java/com/traplaner/travelboardservice/travelBoard/repository/TravelBoardRepository.java
@@ -1,8 +1,8 @@
 package com.traplaner.travelboardservice.travelBoard.repository;
 
-import com.traplaner.travelBoard.entity.TravelBoard;
+import com.traplaner.travelboardservice.travelBoard.entity.TravelBoard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TravelBoardRepository extends JpaRepository<TravelBoard,Long> {
+public interface TravelBoardRepository extends JpaRepository<TravelBoard,Integer> {
 
 }
